feat: reusable CI workflow
Defines a reusable CI workflow based on the standard one used in the majority of Laminas/Mezzio repositories.
(The workflow is deliberately not in `.github/workflows/` to prevent it triggering in this repository.)
Also documents how to consume it via the new `WORKFLOWS.md` file.

We will need to create the tag `ci-v1` after it has been tested and merged, to allow users to pin to a specific known-good version.
Alternately, we can create a branch named `ci-v1` to track the specific workflow and allow making changes over time.
If we do, however, the changes CANNOT require new inputs (though optional inputs would be allowed).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>